requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
idna==3.3
Pillow==9.0.0
pyTelegramBotAPI==4.3.0
requests==2.27.1
urllib3==1.26.8
